backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
redis>=5.0.1
python-socketio>=5.11.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
